django_backend/catalog/serializers.py
from rest_framework import serializers
from .models import Movie

class MovieSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = ['id', 'title', 'description', 'release_year', 'rating', 'image_url', 'created_at', 'updated_at']
